+++ R/list_fields.R
@@ -7,8 +7,8 @@
 #' @seealso \link[countries]{country_info}
 #' @export
 #' @examples
-#' # Run example only if internet connection is available
-#' if (curl::has_internet()){
+#' # Run example only if a connection to the API is available
+#' if (check_countries_api(warnings = FALSE)){
 #'
 #' list_fields()
 #'